requirements.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.2
Pillow==8.1.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
transliterate==1.10.2
